Imports and code refactoring
