+++ src/views/Create/Create.tsx
@@ -1,5 +1,5 @@
-import React, { FC } from 'react'
-import styled from 'styled-components'
+import React, { FC } from 'react';
+import styled from 'styled-components';
 import { Button} from 'components';
 import RequiredFields from './RequiredFields';
 import { fadeIn } from 'App';
